Initial commit
I have this idea kicking around in my head for a cherry MX blue
keyswitch based desk calculator. This is that. The hardware is dead
fucking simple, the firmware is gonna be fun, the mechanical design is
gonna be a bitch and a half (need to find a partner to go in on).

There are problems in this, not everything's connected, one of the
SOT-23-6's needs to be handsoldering pads, the connector for the LCD is
wrong, that connector isn't wired to the microcontroller, I've never
tested this battery config, the USB connector is wrong, the JST
connector is wrong. But overall I like it. Need to maybe consider small
caps on the switch input lines (10pF or so).